\name{ClockstaRG-package}
\alias{ClockstaRG-package}
\alias{ClockstaRG}
\docType{package}
\title{
ClockstaRG: An implementation of ClockstaR for genomic data
}
\description{
CLockstaRG is an implementation of ClockstaR for large genomic data sets. Some of the steps in the algorithm, such as optimising the branch lengths of the trees, and estimating the tree distances can be run in parallel. This implementation also includes an fast clustering algorithm, CLARA, for large data sets.
}
\details{
\tabular{ll}{
Package: \tab ClockstaRG\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-28\cr
License: \tab GPL (<= 2)\cr
}
}
\author{
Sebastian Duchene

Maintainer: Sebastian Duchene <sebastian.duchene@sydney.edu.au>
}
\references{
Duchêne, S, Molak, M., and Ho, SYW. "ClockstaR: choosing the number of relaxed-clock models in molecular phylogenetic analysis." Bioinformatics (2014) 30 (7):1017-1019
}
\keyword{ molecular-clock }
\seealso{
github.com/sebastianduchene/clockstar
github.com/sebastianduchene/nelsi
}
