\name{fill.matrix}
\alias{fill.matrix}
\title{
fill.matirx fills the NA values in the pairwise sBSDmin matrix obtained with fold.sbsd
}
\description{
fill.matirx fills the NA values in the pairwise sBSDmin matrix obtained with fold.sbsd
}
\usage{
fill.matrix(matrix.name, new.matrix.name = NULL)
}
\arguments{
  \item{matrix.name}{
The name of the text file of the matrix obtained with fold.sbsd.
}
  \item{new.matrix.name}{
The file name of the matrix with filled NA values.
}
}
\details{
Please see the tutorial for instructions on how to use this function.
}
\value{
This function does not return a value to the prompt. The output is text file. Please see the tutorial for instructions on how to use this function.
}
\author{
Sebastian Duchene
}
\note{
None. Please see the tutorial for instructions on how to use this function.
}

