\name{make.tree.comps}
\alias{make.tree.comps}
\title{
This function creates a file with the names of the pairwise comparissons for all trees.
}
\description{
This function creates a file with the names of the pairwise comparissons for all trees.
}
\usage{
make.tree.comps(trees.file, tree.comps = "treecomps.txt")
}
\arguments{
  \item{trees.file}{
Name of file with the trees
}
  \item{tree.comps}{
Name of file to write the list of tree comparissons
}
}
\details{
None. See the tutorial for instructions on how to run.
}
\value{
This function does not return values on the command prompt. It creates a text file with the tree comparissons. See the tutorial for instructions on how to run.
}
\references{
NONE
}
\author{
Sebastian Duchene
}
\note{
See the tutorial for instructions on how to run.
}


