\name{run.clara.wk}
\alias{run.clara.wk}
\title{
run.clara is used to run the CLARA algorithm on the MDS of the sBSDmin distances for pairs of trees. For every value of 'k' it estimates the cluster width, 'Wk'.
}
\description{
run.clara is a raper from the CLARA algorithm ,implemented in the package 'cluster. It can be run for a range of values of k.
}
\usage{
run.clara.wk(clus.matrix.name, out.clus.name = "out_clus_wk.txt", k.range = NULL, ...)
}
\arguments{
  \item{clus.matrix.name}{
This is the filename of the MDS of the sBSDmin distances for pairs of trees.
}
  \item{out.clus.name}{
The name of the out put name of the avearage cluster width (Wk) for the values of k in k.range.
}
  \item{k.range}{
The range of 'k' values to calculate Wk.
}
  \item{\dots}{
Other arguments passed to the 'clara' function from pacakge 'cluster'.
 }
}
\details{
Please see the tutorial for instructions.
}
\value{
This function does not return a value to the prompt. It writes a text file with Wk for values of k. 
}
\references{
Kaufman, L., & Rousseeuw, P. J. (2009). Finding groups in data: an introduction to cluster analysis (Vol. 344). John Wiley & Sons.
}
\author{
Sebastian Duchene
}
