\name{scaled_brs}
\alias{scaled_brs}
\title{
Scale branch lengths of tree.
}
\description{
This function is for internal usage. It scales the branch lengths of the trees.
}
\usage{
scaled_brs(tree)
}
\arguments{
  \item{tree}{
  tree with branch lengths. It should be phylo format.
  }
}
\value{
A vector of scaled branch lengths.
}
\author{
Sebastian Duchene
}